requirements.txt
numpy
matplotlib
opencv-python
paho-mqtt
loguru